Revert 95ab285..7b4a2e4
This rolls back to commit 95ab2857d9639f4f55cb5ae215cd04f272cac812.

+++ Matchismo/Model/PlayingCard.m
@@ -54,7 +54,7 @@
 - (int)match:(NSArray *)otherCards
 {
     int score = 0;
-    NSUInteger numOtherCards = [otherCards count];
+    int numOtherCards = [otherCards count];
     
     if (numOtherCards) {
         for (Card *card in otherCards) {
